+++ examples/02_decoding/plot_haxby_grid_search.py
@@ -97,8 +97,7 @@ anova_svc = Pipeline([('anova', feature_selection), ('svc', svc)])
 anova_svc.fit(X, y)
 y_pred = anova_svc.predict(X)
 
-from sklearn.cross_validation import LeaveOneLabelOut, cross_val_score
-cv = LeaveOneLabelOut(session[session < 10])
+from sklearn.model_selection import cross_val_score
 
 k_range = [10, 15, 30, 50, 150, 300, 500, 1000, 1500, 3000, 5000]
 cv_scores = []
@@ -118,7 +117,7 @@ for k in k_range:
 ###########################################################################
 # Nested cross-validation
 # -------------------------
-from sklearn.grid_search import GridSearchCV
+from sklearn.model_selection import GridSearchCV
 # We are going to tune the parameter 'k' of the step called 'anova' in
 # the pipeline. Thus we need to address it as 'anova__k'.