+++ examples/02_decoding/plot_oasis_vbm_space_net.py
@@ -26,7 +26,7 @@ gm_imgs = np.array(dataset_files.gray_matter_maps)
 
 # Split data into training set and test set
 from sklearn.utils import check_random_state
-from sklearn.cross_validation import train_test_split
+from sklearn.model_selection import train_test_split
 rng = check_random_state(42)
 gm_imgs_train, gm_imgs_test, age_train, age_test = train_test_split(
     gm_imgs, age, train_size=.6, random_state=rng)