+++ nilearn/plotting/tests/test_surf_plotting.py
@@ -16,11 +16,6 @@ from nilearn.surface.tests.test_surface import _generate_surf
 
 
 def test_plot_surf():
-    # Axes3DSubplot has no attribute 'plot_trisurf' for older versions of
-    # matplotlib
-    if LooseVersion(matplotlib.__version__) <= LooseVersion('1.3.1'):
-        raise SkipTest
-
     mesh = _generate_surf()
     rng = np.random.RandomState(0)
     bg = rng.randn(mesh[0].shape[0], )
@@ -38,15 +33,14 @@ def test_plot_surf():
     plot_surf(mesh, bg_map=bg, view='medial')
     plot_surf(mesh, bg_map=bg, hemi='right', view='medial')
 
+    # Plot with colorbar
+    plot_surf(mesh, bg_map=bg, colorbar=True)
+
     # Save execution time and memory
     plt.close()
 
 
 def test_plot_surf_error():
-    # Axes3DSubplot has no attribute 'plot_trisurf' for older versions of
-    # matplotlib
-    if LooseVersion(matplotlib.__version__) <= LooseVersion('1.3.1'):
-        raise SkipTest
     mesh = _generate_surf()
     rng = np.random.RandomState(0)
 
@@ -74,11 +68,6 @@ def test_plot_surf_error():
 
 
 def test_plot_surf_stat_map():
-    # Axes3DSubplot has no attribute 'plot_trisurf' for older versions of
-    # matplotlib
-    if LooseVersion(matplotlib.__version__) <= LooseVersion('1.3.1'):
-        raise SkipTest
-
     mesh = _generate_surf()
     rng = np.random.RandomState(0)
     bg = rng.randn(mesh[0].shape[0], )
@@ -86,38 +75,61 @@ def test_plot_surf_stat_map():
 
     # Plot mesh with stat map
     plot_surf_stat_map(mesh, stat_map=data)
+    plot_surf_stat_map(mesh, stat_map=data, colorbar=True)
     plot_surf_stat_map(mesh, stat_map=data, alpha=1)
 
     # Plot mesh with background and stat map
     plot_surf_stat_map(mesh, stat_map=data, bg_map=bg)
     plot_surf_stat_map(mesh, stat_map=data, bg_map=bg,
-                       bg_on_stat=True, darkness=0.5)
+                       bg_on_data=True, darkness=0.5)
+    plot_surf_stat_map(mesh, stat_map=data, bg_map=bg, colorbar=True,
+                       bg_on_data=True, darkness=0.5)
 
     # Apply threshold
     plot_surf_stat_map(mesh, stat_map=data, bg_map=bg,
-                       bg_on_stat=True, darkness=0.5,
+                       bg_on_data=True, darkness=0.5,
+                       threshold=0.3)
+    plot_surf_stat_map(mesh, stat_map=data, bg_map=bg, colorbar=True,
+                       bg_on_data=True, darkness=0.5,
                        threshold=0.3)
 
     # Change vmax
     plot_surf_stat_map(mesh, stat_map=data, vmax=5)
+    plot_surf_stat_map(mesh, stat_map=data, vmax=5, colorbar=True)
 
     # Change colormap
     plot_surf_stat_map(mesh, stat_map=data, cmap='cubehelix')
+    plot_surf_stat_map(mesh, stat_map=data, cmap='cubehelix', colorbar=True)
 
     # Plot to axes
     axes = plt.subplots(ncols=2, subplot_kw={'projection': '3d'})[1]
     for ax in axes.flatten():
         plot_surf_stat_map(mesh, stat_map=data, ax=ax)
+    axes = plt.subplots(ncols=2, subplot_kw={'projection': '3d'})[1]
+    for ax in axes.flatten():
+        plot_surf_stat_map(mesh, stat_map=data, ax=ax, colorbar=True)
 
+    fig = plot_surf_stat_map(mesh, stat_map=data, colorbar=False)
+    assert len(fig.axes) == 1
+    # symmetric_cbar
+    fig = plot_surf_stat_map(
+        mesh, stat_map=data, colorbar=True, symmetric_cbar=True)
+    assert len(fig.axes) == 2
+    yticklabels = fig.axes[1].get_yticklabels()
+    first, last = yticklabels[0].get_text(), yticklabels[-1].get_text()
+    assert float(first) == - float(last)
+    # no symmetric_cbar
+    fig = plot_surf_stat_map(
+        mesh, stat_map=data, colorbar=True, symmetric_cbar=False)
+    assert len(fig.axes) == 2
+    yticklabels = fig.axes[1].get_yticklabels()
+    first, last = yticklabels[0].get_text(), yticklabels[-1].get_text()
+    assert float(first) != - float(last)
     # Save execution time and memory
     plt.close()
 
 
 def test_plot_surf_stat_map_error():
-    # Axes3DSubplot has no attribute 'plot_trisurf' for older versions of
-    # matplotlib
-    if LooseVersion(matplotlib.__version__) <= LooseVersion('1.3.1'):
-        raise SkipTest
     mesh = _generate_surf()
     rng = np.random.RandomState(0)
     data = 10 * rng.randn(mesh[0].shape[0], )
@@ -140,10 +152,6 @@ def test_plot_surf_stat_map_error():
 
 
 def test_plot_surf_roi():
-    # Axes3DSubplot has no attribute 'plot_trisurf' for older versions of
-    # matplotlib
-    if LooseVersion(matplotlib.__version__) <= LooseVersion('1.3.1'):
-        raise SkipTest
     mesh = _generate_surf()
     rng = np.random.RandomState(0)
     roi1 = rng.randint(0, mesh[0].shape[0], size=5)
@@ -152,12 +160,15 @@ def test_plot_surf_roi():
 
     # plot roi
     plot_surf_roi(mesh, roi_map=roi1)
+    plot_surf_roi(mesh, roi_map=roi1, colorbar=True)
 
     # plot parcellation
     plot_surf_roi(mesh, roi_map=parcellation)
+    plot_surf_roi(mesh, roi_map=parcellation, colorbar=True)
 
     # plot roi list
     plot_surf_roi(mesh, roi_map=[roi1, roi2])
+    plot_surf_roi(mesh, roi_map=[roi1, roi2], colorbar=True)
 
     # plot to axes
     plot_surf_roi(mesh, roi_map=roi1, ax=None, figure=plt.gcf())
@@ -166,6 +177,9 @@ def test_plot_surf_roi():
     with tempfile.NamedTemporaryFile() as tmp_file:
         plot_surf_roi(mesh, roi_map=roi1, ax=plt.gca(), figure=None,
                       output_file=tmp_file.name)
+    with tempfile.NamedTemporaryFile() as tmp_file:
+        plot_surf_roi(mesh, roi_map=roi1, ax=plt.gca(), figure=None,
+                      output_file=tmp_file.name, colorbar=True)
 
     # Save execution time and memory
     plt.close()